requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==4.4.2
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1
django-js-asset==2.0.0
django-modelcluster==6.0
django-otp==1.1.6
django-taggit==3.1.0
django-tinymce==3.5.0
filelock==3.12.0
gTTS==2.5.1
gunicorn==22.0.0
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
moviepy==1.0.3
mutagen==1.47.0
numpy==1.26.4
packaging==24.0
pillow==10.3.0
platformdirs==3.2.0
proglog==0.1.10
psycopg2-binary==2.9.5
pycryptodomex==3.20.0
pypdf==4.2.0
PyPDF2==3.0.1
pypng==0.20220715.0
pytesseract==0.3.10
python-decouple==3.6
pyttsx3==2.90
pytz==2023.3
qrcode==7.4.2
requests==2.31.0
sqlparse==0.5.0
tesseract==0.1.3
tqdm==4.66.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.2.1
virtualenv==20.22.0
virtualenv-clone==0.5.7
websockets==12.0
whitenoise==6.6.0
yt-dlp==2024.4.9
